+++ src/main/java/com/deca/NoticeBoardWebSite/domain/PostData.java
@@ -8,7 +8,7 @@ import java.util.*;
 // getter, setter 추가, 문자열로 자동변환
 public class PostData {
     /** 게시물 id, 생성된 순서대로 1부터 증가 */
-    private Integer id;
+    private Long id;
 
     /** 게시물 제목 */
     private String title;
@@ -23,13 +23,13 @@ public class PostData {
     private String type;
 
     /** 게시물의 조회수 */
-    private Integer viewCount;
+    private Long viewCount;
 
     /** 게시물의 좋아요 수 */
-    private Integer like;
+    private Long like;
 
     /** 게시물의 싫어요 수 */
-    private Integer disLike;
+    private Long disLike;
 
     /**
      * 기본 생성자
@@ -38,26 +38,26 @@ public class PostData {
      * @param type 게시판 종류
      */
     public PostData(String title, String content, String type) {
-        this.id = 0; // save 에서 set
+        this.id = 0L; // save 에서 set
         this.title = title;
         this.content = content;
         this.time = ""; // save 에서 set
         this.type = type;
-        this.viewCount = 0;
-        this.like = 0;
-        this.disLike = 0;
+        this.viewCount = 0L;
+        this.like = 0L;
+        this.disLike = 0L;
     }
 
     /** 매개변수 없는 생성자, 기본값들이 들어감 */
     public PostData() {
-        this.id = 0; // save 에서 set
+        this.id = 0L; // save 에서 set
         this.title = "new Title";
         this.content = "empty content";
         this.time = ""; // save 에서 set
         this.type = "free";
-        this.viewCount = 0;
-        this.like = 0;
-        this.disLike = 0;
+        this.viewCount = 0L;
+        this.like = 0L;
+        this.disLike = 0L;
     }
 
     // 추가할 예정: 작성자, 댓글
